An extension point for code analyzers that are run as part of the build process.
Analyzers can be enabled individually, and it's possible to define a custom
severity for the markers. Preferences can also be set per project. The analysis
is implemented as a separate plug-in so that it can be used from the SBT
builder.

For development and running inside Eclipse, we now create a separate
compiler-plugin jar that then loads the analyzers from the running
Eclipse's classpath. The provided builder and launch config contain
everything to make things work.